Decompose into Modules
[Post Merge of PR #106]

Generalize Configuration, Fix Token Handling, RateLimiter Class,
and Fix Message Splitting.
